website build update added syntax highlight theme twilight

+++ web/org-split-body-toc.sh
@@ -3,14 +3,24 @@
 # small shell script to collate together the web pages
 # extracts the <body> part from the org-mode rendered HTML file
 # then formats it as a template to be shown by the index.php
+#
+# March 2010 by Jaromil
 
+if [ -z $1 ]; then
+    echo "usage: org-split-body-toc.sh orgmode.html"
+    echo "this tool splits an orgmode in two parts: <TOC> and <body>"
+    echo "it also eliminates headers and footers."
+    exit 1
+fi
 
-tmpfile="/tmp/collate-dyne-web-`date +%s`"
+tmpfile="/tmp/org-split-body-toc-`date +%s`"
 rm -f $tmpfile
 
-rm -f body.html toc.html
+orgfile=${1}
 
-cat index.html | \
+orgname=`echo $orgfile|cut -d. -f1`
+
+cat $orgfile | \
     awk '
 BEGIN       { body_found=0; title_found=0; }
 /<body>/             { body_found=1; next; }
@@ -27,7 +37,7 @@ if(body_found==1)
 ' > $tmpfile
 
 # separate the TOC
-echo "<div id=\"table-of-contents\">" > toc.html
+echo "<div id=\"table-of-contents\">" > ${orgname}-toc.html
 cat $tmpfile | \
     awk '
 BEGIN { toc_found=0; div_count=0; }
@@ -36,7 +46,7 @@ BEGIN { toc_found=0; div_count=0; }
   if(toc_found==1) print $0;
 }
 /^<\/div>/ { div_count++; }
-' >> toc.html
+' >> ${orgname}-toc.html
 
 # separate the BODY
 cat $tmpfile | \
@@ -44,8 +54,11 @@ cat $tmpfile | \
 BEGIN { div_count=0; }
 { if(div_count>1) print $0; }
 /^<\/div>/ { div_count++; }
-' > body.html
-echo "</div>" >> body.html
+' > ${orgname}-body.html
+echo "</div>" >> ${orgname}-body.html
 
-# deleted by the cronscript on server
-rm -f index.html
+# delete the temporary file
+rm -f $tmpfile
+
+echo "org-mode html $orgfile splitted in ${orgname}-toc.html and ${orgname}-body.html"
+
